requirements.txt
Flask==3.0.3
flask-sqlalchemy==3.1.1
gunicorn==22.0.0
yt-dlp==2024.12.13
requests==2.31.0
google-api-python-client==2.140.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
Pillow==10.4.0
python-dotenv==1.0.1
apscheduler==3.10.4